Fixed the guaranteed-dictionary-update bug
It wasn't actually guaranteed, it turns out.  I was iterating over
dictionary keys, which is done in arbitrary order.  The result was that
in different executions of the program, the corpus was generated
differently, so the hashes differed, and the language had to be updated.
Sorting the keys before adding them to the list of number-words fixed
the problem.