Rewrote NumberParser.parse_all_numbers()
Its control flow was confusing before; now it's much more
straightforward.  We make a string representing classes of words, split
that by a regular expression for number words, then parse each number
and build up our return string and list.  It works just as well as the
previous method, is a bit shorter, and I feel that it's clearer as well.