A bit of cleanup in the README
It actually does do the things claimed in the first sentence, so don't
use the future tense.

There's no need to tell the user the details of intermediate steps in
generating the language.  All the user needs to know is that it uses the
network to update the language model when it starts.